+++ feature/fido2/src/androidMain/kotlin/com/chimali/fido2/util/logging/AndroidLogDirectoryProvider.kt
@@ -1,0 +1,11 @@
+package com.chimali.fido2.util.logging
+
+import android.content.Context
+import okio.Path
+import okio.Path.Companion.toOkioPath
+
+class AndroidLogDirectoryProvider(private val context: Context) : LogDirectoryProvider {
+    override fun getLogDirectory(): Path {
+        return context.filesDir.resolve("logs").toOkioPath()
+    }
+}