+++ feature/fido2/src/androidMain/kotlin/com/chimali/fido2/util/logging/PlatformLock.kt
@@ -1,0 +1,15 @@
+package com.chimali.fido2.util.logging
+
+import java.util.concurrent.locks.ReentrantLock
+
+actual class PlatformLock actual constructor() {
+    private val reentrantLock = ReentrantLock()
+
+    actual fun lock() {
+        reentrantLock.lock()
+    }
+
+    actual fun unlock() {
+        reentrantLock.unlock()
+    }
+}