+++ feature/fido2/src/commonMain/kotlin/com/chimali/fido2/util/logging/LogDirectoryProvider.kt
@@ -1,0 +1,7 @@
+package com.chimali.fido2.util.logging
+
+import okio.Path
+
+interface LogDirectoryProvider {
+    fun getLogDirectory(): Path
+}