+++ feature/fido2/src/commonMain/kotlin/com/chimali/fido2/util/logging/PlatformLock.kt
@@ -1,0 +1,15 @@
+package com.chimali.fido2.util.logging
+
+expect class PlatformLock() {
+    fun lock()
+    fun unlock()
+}
+
+inline fun <T> PlatformLock.withLock(block: () -> T): T {
+    lock()
+    try {
+        return block()
+    } finally {
+        unlock()
+    }
+}