+++ feature/fido2/src/iosMain/kotlin/com/chimali/fido2/util/logging/PlatformLock.kt
@@ -1,0 +1,15 @@
+package com.chimali.fido2.util.logging
+
+import platform.Foundation.NSRecursiveLock
+
+actual class PlatformLock actual constructor() {
+    private val nsLock = NSRecursiveLock()
+
+    actual fun lock() {
+        nsLock.lock()
+    }
+
+    actual fun unlock() {
+        nsLock.unlock()
+    }
+}